chore: enable Prettier formatting
This commit enables Prettier formatting to maintain a consistent code style
across the entire codebase. It's now a community standard and easy to use with
tools like Husky, which set up Git hooks to run formatting before actions such
as `git commit`.

There are no functional changes in this commit, only stylish updates to the formatting.